refactor: User 기능에서 Auth 기능 분리
