+++ src/main/java/roomescape/auth/domain/dto/TokenRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.domain.dto;
 
 public record TokenRequestDto(String email, String password) {
 }
