+++ src/main/java/roomescape/user/controller/UserController.java
@@ -1,15 +1,7 @@
 package roomescape.user.controller;
 
-import java.time.Duration;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseCookie;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.TokenRequestDto;
-import roomescape.auth.TokenResponseDto;
 import roomescape.user.service.UserService;
 
 @RestController
@@ -20,23 +12,5 @@ public class UserController {
 
     public UserController(UserService userService) {
         this.userService = userService;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(@RequestBody TokenRequestDto tokenRequestDto) {
-
-        TokenResponseDto tokenResponseDto = userService.login(tokenRequestDto);
-        
-        ResponseCookie cookie = ResponseCookie
-                .from("SESSION", tokenResponseDto.accessToken())
-                .domain("localhost")
-                .path("/")
-                .httpOnly(true)
-                .secure(false)
-                .maxAge(Duration.ofDays(30))
-                .sameSite("Strict")
-                .build();
-
-        return ResponseEntity.ok().header(HttpHeaders.SET_COOKIE, cookie.toString()).build();
     }
 }
