+++ src/main/resources/templates/admin/index.html
@@ -1,60 +1,58 @@
 <!DOCTYPE html>
 <html lang="ko">
 <head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>방탈출 어드민</title>
-  <!-- Bootstrap CSS -->
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-  <link rel="stylesheet" href="/css/style.css">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>방탈출 어드민</title>
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+    <link rel="stylesheet" href="/css/style.css">
 </head>
 <body>
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
-  <a class="navbar-brand" href="/admin">
-    <img src="/image/admin-logo.png" alt="LOGO" style="height: 40px;">
-  </a>
-  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
-          aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-    <span class="navbar-toggler-icon"></span>
-  </button>
+    <a class="navbar-brand" href="/admin">
+        <img src="/image/admin-logo.png" alt="LOGO" style="height: 40px;">
+    </a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+    </button>
 
-  <div class="collapse navbar-collapse" id="navbarSupportedContent">
-    <ul class="navbar-nav ml-auto">
-      <li class="nav-item">
-        <a class="nav-link" href="/admin/reservation">Reservation</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="/admin/theme">Theme</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="/admin/time">Time</a>
-      </li>
-      <!--      TODO: [4단계] 사용자 로그인 - 로그인 기능 구현 후 활성화-->
-      <!--      <li class="nav-item">-->
-      <!--        <a class="nav-link" href="/login">Login</a>-->
-      <!--      </li>-->
-      <!--      <li class="nav-item dropdown">-->
-      <!--        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown"-->
-      <!--           aria-haspopup="true" aria-expanded="false">-->
-      <!--          <img class="profile-image" src="/image/default-profile.png" alt="Profile">-->
-      <!--          <span id="profile-name">Profile</span> &lt;!&ndash; 프로필 이름을 넣을 span 추가 &ndash;&gt;-->
-      <!--        </a>-->
-      <!--        <div class="dropdown-menu" aria-labelledby="navbarDropdown">-->
-      <!--          <a class="dropdown-item" href="/reservation-mine">My Reservation</a>-->
-      <!--          <div class="dropdown-divider"></div>-->
-      <!--          <a class="dropdown-item" href="#" id="logout-btn">Logout</a>-->
-      <!--        </div>-->
-      <!--      </li>-->
-    </ul>
-  </div>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav ml-auto">
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/reservation">Reservation</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/theme">Theme</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/admin/time">Time</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/login">Login</a>
+            </li>
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown"
+                   aria-haspopup="true" aria-expanded="false">
+                    <img class="profile-image" src="/image/default-profile.png" alt="Profile">
+                    <span id="profile-name">Profile</span> <!-- 프로필 이름을 넣을 span 추가 -->
+                </a>
+                <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                    <a class="dropdown-item" href="/reservation-mine">My Reservation</a>
+                    <div class="dropdown-divider"></div>
+                    <a class="dropdown-item" href="#" id="logout-btn">Logout</a>
+                </div>
+            </li>
+        </ul>
+    </div>
 </nav>
 
 <div class="content-container">
-  <h2 class="content-container-title">방탈출 어드민</h2>
+    <h2 class="content-container-title">방탈출 어드민</h2>
 </div>
 
-<!--TODO: [4단계] 사용자 로그인 - 로그인 기능 구현 후 활성화-->
-<!--<script src="/js/user-scripts.js"></script>-->
+<script src="/js/user-scripts.js"></script>
 </body>
 </html>
