+++ src/test/java/roomescape/auth/controller/AuthControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.user.controller;
+package roomescape.auth.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -10,12 +10,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import roomescape.auth.TokenRequestDto;
+import roomescape.auth.domain.dto.TokenRequestDto;
 import roomescape.user.fixture.UserFixture;
 import roomescape.user.service.UserService;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-class UserControllerTest {
+class AuthControllerTest {
 
     private static final String EMAIL = "user@email.com";
     private static final String PASSWORD = "password";
@@ -53,7 +53,7 @@ class UserControllerTest {
                     .given().log().all()
                     .contentType(ContentType.JSON)
                     .body(requestDto)
-                    .when().post("/users/login")
+                    .when().post("/login")
                     .then().log().all()
                     .statusCode(HttpStatus.OK.value())
                     .extract().response();
@@ -72,7 +72,7 @@ class UserControllerTest {
                     .given().log().all()
                     .contentType(ContentType.JSON)
                     .body(requestDto)
-                    .when().post("/users/login")
+                    .when().post("/login")
                     .then().log().all()
                     .statusCode(HttpStatus.NOT_FOUND.value())
                     .extract().response();
